chore: refactor?
