--- .prettierrc
@@ -1,8 +1,0 @@
-{
-	"semi": true,
-	"trailingComma": "all",
-	"singleQuote": true,
-	"printWidth": 120,
-	"useTabs": true,
-	"tabWidth": 2
-}
