+++ prettier.config.mjs
@@ -1,0 +1,10 @@
+const config = {
+	semi: true,
+	trailingComma: 'all',
+	singleQuote: true,
+	printWidth: 120,
+	useTabs: true,
+	tabWidth: 2,
+};
+
+export default config;
