+++ src/cron/job.ts
@@ -8,9 +8,9 @@ import { app } from '@/app';
 import { IInstallation } from '@/entities/installation';
 import { IScoreLog } from '@/entities/scoreLog';
 import { DateUtil } from '@/lib/date';
-import * as configService from '@/lib/services/configService';
-import * as scoreboardService from '@/lib/services/scoreboardService';
-import * as slackService from '@/lib/services/slack';
+import * as configService from '@/services/configService';
+import * as scoreboardService from '@/services/scoreboardService';
+import * as slackService from '@/services/slack';
 import { withNamespace } from '@/logger';
 import config from '@config';
 import { ChatPostMessageArguments } from '@slack/web-api';
