+++ src/lib/services/scorekeeperService.ts
@@ -4,14 +4,14 @@ import { app } from '@/app';
 import { config } from '@/config';
 import { IScoreLog } from '@/entities/scoreLog';
 import { IUser } from '@/entities/user';
-import * as adminService from '@/lib/services/adminService';
-import * as botTokenService from '@/lib/services/botTokenService';
-import * as configService from '@/lib/services/configService';
-import * as databaseService from '@/lib/services/databaseService';
-import { eventBus } from '@/lib/services/eventBus';
-import * as installService from '@/lib/services/installService';
-import * as scoreLogService from '@/lib/services/scoreLogService';
-import * as userService from '@/lib/services/userService';
+import * as adminService from '@/services/adminService';
+import * as botTokenService from '@/services/botTokenService';
+import * as configService from '@/services/configService';
+import * as databaseService from '@/services/databaseService';
+import { eventBus } from '@/services/eventBus';
+import * as installService from '@/services/installService';
+import * as scoreLogService from '@/services/scoreLogService';
+import * as userService from '@/services/userService';
 import { PPSpamEvent, PPSpamEventName } from '@/lib/types/Events';
 import { withNamespace } from '@/logger';
 import { ChatPostMessageArguments } from '@slack/web-api';
