+++ src/monthlyScoreboardCron.ts
@@ -1,4 +1,4 @@
-import * as installService from '@/lib/services/installService';
+import * as installService from '@/services/installService';
 import { withNamespace } from '@/logger';
 import { createCronJob } from './cron';
